fix: Update format specifiers for elapsed time display to use long long

+++ src/cards/information/calendar/calendar.hpp
@@ -125,7 +125,7 @@ namespace rouen::cards
             auto minutes = elapsed_seconds / 60;
             auto seconds = elapsed_seconds % 60;
             
-            ImGui::Text("Last refresh: %ld:%02ld ago", minutes, seconds);
+            ImGui::Text("Last refresh: %lld:%02lld ago", static_cast<long long>(minutes), static_cast<long long>(seconds));
             ImGui::SameLine();
             
             if (ImGui::Button("Refresh Now")) {
